pyproject.toml
[project]
name = "TradingTools"
version = "0.1.0"
description = "A basic Flask app for using custom trading tools and visualising price data in Lightweight Charts from Trading View"
authors = [
    { name = "PoivronJaune", email = "poivronjaune@gmail.com" }
]
dependencies = [
    "Flask>=2.0.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.1.70",
    "pandas>=1.0" 
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]


[tool.setuptools]
packages = ["webchart"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
